requirements.txt
dj-database-url==0.4.2
Django==1.11
django-embed-video==1.1.0
django-registration-redux==1.4
easy-thumbnails==2.4.1
gunicorn==19.7.1
olefile==0.44
Pillow==4.1.0
# pkg-resources==0.0.0
psycopg2==2.7.1
pytz==2017.2
requests==2.13.0
